Refactor reqlog.h and event type to be an enum
Refactor reqlog functions into its own reqlog.h file out from comdb2.h
Have event type to be an enum
